+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 2026-04-25 13:46 EST
+- Improve plugin chooser recommendations with instrument/bus context filters and five-star ratings on web and iOS.
+
 ## 2026-04-25 12:22 EST
 - Replace repeated detail-card table subtitles with compact icon labels on web and iOS.